@@ -1,11 +1,11 @@
 name: transformer_page_view
 description: PageTransformer for flutter
-version: 0.1.6
+version: 0.2.0
 author: JZoom <jzoom8112@gmail.com>
 homepage:  https://github.com/jzoom/transformer_page_view
 
 environment:
-  sdk: ">=2.0.0-dev.50.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
